// client/src/components/checkout/types/checkout.client.ts
export type BuyerState = {
  fullName: string;
  email: string;
  phone: string;
  city: string;
  district: string;
  ward: string;
  street: string;
  notes: string;
  shipToDifferent: boolean;
};

export type AltState = {
  altFullName: string;
  altPhone: string;
  altCity: string;
  altDistrict: string;
  altWard: string;
  altStreet: string;
};